Refactoring
Moved PluginController into plugin.py to mitigate a cyclic import issue.

Replaced stt_tts.py's Speech class with Input, which uses static methods instead.
